+++ src/BDSDetectorConstruction.cc
@@ -159,7 +159,7 @@ void BDSDetectorConstruction::UpdateSamplerDiameterAndCountSamplers()
       G4double curvilinearRadiusBends = (0.9 / maxBendingRatio)*CLHEP::m; // 90% of theoretical maximum radius
 
       // check it's smaller - the user may have already specified a smaller sampler diameter
-      // and htat should take precedence
+      // and that should take precedence
       if (curvilinearRadiusBends < curvilinearRadius)
 	{
 	  curvilinearRadius = curvilinearRadiusBends;
@@ -485,7 +485,7 @@ G4VPhysicalVolume* BDSDetectorConstruction::BuildWorld()
   const auto& blMain = acceleratorModel->BeamlineSetMain();
   blMain.GetExtentGlobals(extents);
 
-  // check optional placement beam line (likevector of placements)
+  // check optional placement beam line (like vector of placements)
   BDSBeamline* plBeamline = acceleratorModel->PlacementBeamline();
   if (plBeamline) // optional - may be nullptr
     {extents.push_back(plBeamline->GetExtentGlobal());}
@@ -718,7 +718,7 @@ void BDSDetectorConstruction::PlaceBeamlineInWorld(BDSBeamline*          beamlin
 	  
 	  BDSPhysicalVolumeInfoRegistry::Instance()->RegisterInfo(pv, theinfo, true);
         }
-      i++; // for increuemental copy numbers
+      i++; // for incremental copy numbers
     }
 }
 
@@ -761,6 +761,8 @@ G4Transform3D BDSDetectorConstruction::CreatePlacementTransform(const GMAD::Plac
   // s and use as local 'z' in the transform.
   if (!placement.referenceElement.empty())
     {// scenario 3
+      if (!beamLine)
+        {throw BDSException(__METHOD_NAME__, "no valid beam line yet placement w.r.t. a beam line.");}
       const BDSBeamlineElement* element = beamLine->GetElement(placement.referenceElement,
 							 placement.referenceElementNumber);
       if (!element)