+++ src/BDSOutputROOTEventSampler.cc
@@ -79,6 +79,7 @@ void BDSOutputROOTEventSampler<U>::Fill(const BDSHitSampler* hit,
   xp.push_back((U &&) (hit->coords.xp / CLHEP::radian));
   yp.push_back((U &&) (hit->coords.yp / CLHEP::radian));
   zp.push_back((U &&) (hit->coords.zp / CLHEP::radian));
+  p.push_back((U &&) (hit->momentum / CLHEP::GeV));
   T.push_back((U &&) (hit->coords.T / CLHEP::ns));
   
   modelID = hit->beamlineIndex;
@@ -89,7 +90,6 @@ void BDSOutputROOTEventSampler<U>::Fill(const BDSHitSampler* hit,
   trackID.push_back(hit->trackID);
   turnNumber.push_back(hit->turnsTaken);
 
-  // require mass to calculate kinetic energ
   if (storeMass)
     {mass.push_back((double)(hit->mass / CLHEP::GeV));}
 
@@ -111,14 +111,18 @@ void BDSOutputROOTEventSampler<U>::Fill(const BDSHitSampler* hit,
 
 template <class U>
 void BDSOutputROOTEventSampler<U>::Fill(const BDSParticleCoordsFull& coords,
-					const G4double chargeIn,
-					const G4int    pdgID,
-					const G4int    turnsTaken,
-					const G4int    beamlineIndex,
-					const G4int    nElectronsIn,
-					const G4double massIn,
-					const G4double rigidityIn,
-					G4bool fillIon)
+					G4double       momentumIn,
+					G4double chargeIn,
+					G4int    pdgID,
+					G4int    turnsTaken,
+					G4int    beamlineIndex,
+					G4int    nElectronsIn,
+					G4double massIn,
+					G4double rigidityIn,
+					G4bool  fillIon,
+					G4bool* isIonIn,
+					G4int*  ionAIn,
+					G4int*  ionZIn)
 {
   trackID.push_back(n); // we assume multiple primaries are linearly increasing in track number
   n++;
@@ -129,6 +133,7 @@ void BDSOutputROOTEventSampler<U>::Fill(const BDSParticleCoordsFull& coords,
   xp.push_back((U &&) (coords.xp / CLHEP::radian));
   yp.push_back((U &&) (coords.yp / CLHEP::radian));
   zp.push_back((U &&) (coords.zp / CLHEP::radian));
+  p.push_back((U &&) (momentumIn / CLHEP::GeV));
   T.push_back((U &&) (coords.T / CLHEP::ns));
   weight.push_back((const U &) coords.weight);
   partID.push_back(pdgID);
@@ -144,7 +149,13 @@ void BDSOutputROOTEventSampler<U>::Fill(const BDSParticleCoordsFull& coords,
   nElectrons.push_back((int)nElectronsIn);
   kineticEnergy.push_back((double)(coords.totalEnergy - massIn) / CLHEP::GeV);
   FillPolarCoords(coords);
-  if (fillIon)
+  if (isIonIn && ionAIn && ionZIn)
+    {
+      isIon.push_back((int) *isIonIn);
+      ionA.push_back((int) *ionAIn);
+      ionZ.push_back((int) *ionZIn);
+    }
+  else if (fillIon)
     {FillIon();}
 }
 
@@ -157,29 +168,31 @@ void BDSOutputROOTEventSampler<U>::FillPolarCoords(const BDSParticleCoordsFull& 
   double ypCoord = coords.yp;
   double zpCoord = coords.zp;
 
-  // we have to tolerate possible sqrt errors here
-  double rValue = std::sqrt(std::pow(xCoord, 2) + std::pow(yCoord, 2));
-  if (!std::isnormal(rValue))
+  // nans or infinite numbers can be set to -1
+  auto isntSafe = [](double a){return std::isnan(a) || std::isinf(a);};
+
+  double rValue = std::hypot(xCoord, yCoord);
+  if (isntSafe(rValue))
     {rValue = 0;}
   r.push_back(static_cast<U>(rValue));
   
-  double rpValue = std::sqrt(std::pow(xpCoord, 2) + std::pow(ypCoord, 2));
-  if (!std::isnormal(rpValue))
+  double rpValue = std::hypot(xpCoord, ypCoord);
+  if (isntSafe(rpValue))
     {rpValue = 0;}
   rp.push_back(static_cast<U>(rpValue));
 
   double thetapValue = std::atan2(rpValue, zpCoord);
-  if (!std::isnormal(thetapValue))
+  if (isntSafe(thetapValue))
     {thetapValue = -1;}
   theta.push_back(thetapValue);
 
   double phiValue = std::atan2(xCoord, yCoord);
-  if (!std::isnormal(phiValue))
+  if (isntSafe(phiValue))
     {phiValue = -1;}
   phi.push_back(static_cast<U>(phiValue));
 
   double phipValue = std::atan2(xpCoord, ypCoord);
-  if (!std::isnormal(phipValue))
+  if (isntSafe(phipValue))
     {phipValue = -1;}
   phip.push_back(static_cast<U>(phipValue));
 }
@@ -191,6 +204,7 @@ void BDSOutputROOTEventSampler<U>::Fill(const BDSPrimaryVertexInformationV* vert
   for (const auto& vertexInfo : vertexInfos->vertices)
     {
       Fill(vertexInfo.primaryVertex.local,
+	   vertexInfo.momentum,
 	   vertexInfo.charge,
 	   vertexInfo.pdgID,
 	   turnsTaken,
@@ -202,10 +216,6 @@ void BDSOutputROOTEventSampler<U>::Fill(const BDSPrimaryVertexInformationV* vert
     }
   FillIon();
 }
-
-//#else
-//void BDSOutputROOTEventSampler::SetBranchAddress(TTree *)
-//{}
 #endif
 
 template <class U>
@@ -221,6 +231,7 @@ void BDSOutputROOTEventSampler<U>::Fill(const BDSOutputROOTEventSampler<U>* othe
   xp     = other->xp;
   yp     = other->yp;
   zp     = other->zp;
+  p      = other->p;
   T      = other->T;
 
   weight     = other->weight;
@@ -261,6 +272,7 @@ template <class U> void BDSOutputROOTEventSampler<U>::Flush()
   xp.clear();
   yp.clear();
   zp.clear();
+  p.clear();
   T.clear();
   weight.clear();
   partID.clear();