+++ src/BDSPrimaryVertexInformation.cc
@@ -28,6 +28,7 @@ along with BDSIM.  If not, see <http://www.gnu.org/licenses/>.
 BDSPrimaryVertexInformation:: BDSPrimaryVertexInformation(const BDSParticleCoordsFullGlobal& primaryVertexIn,
 							  const BDSParticleDefinition*       particle):
   primaryVertex(primaryVertexIn),
+  momentum(particle->Momentum()),
   charge(particle->Charge()),
   rigidity(particle->BRho()),
   mass(particle->Mass()),
@@ -35,13 +36,15 @@ BDSPrimaryVertexInformation:: BDSPrimaryVertexInformation(const BDSParticleCoord
   nElectrons(particle->NElectrons())
 {;}
 
-BDSPrimaryVertexInformation:: BDSPrimaryVertexInformation(const BDSParticleCoordsFullGlobal& primaryVertexIn,
-							  const G4double chargeIn,
-							  const G4double rigidityIn,
-							  const G4double massIn,
-							  const G4int    pdgIDIn,
-							  const G4int    nElectronsIn):
+BDSPrimaryVertexInformation::BDSPrimaryVertexInformation(const BDSParticleCoordsFullGlobal& primaryVertexIn,
+							  G4double       momentumIn,
+							  G4double chargeIn,
+							  G4double rigidityIn,
+							  G4double massIn,
+							  G4int    pdgIDIn,
+							  G4int    nElectronsIn):
   primaryVertex(primaryVertexIn),
+  momentum(momentumIn),
   charge(chargeIn),
   rigidity(rigidityIn),
   mass(massIn),